+++ resources/views/admin/commissions/create_edit_commission_class.blade.php
@@ -107,6 +107,15 @@
         </div>
     @endif
 
+    @if (config('aldebaran.commissions.payment_processors.stripe.integration.enabled') || config('aldebaran.commissions.payment_processors.paypal.integration.enabled'))
+        <h2>Invoice Information</h2>
+        <p>
+            This will be used to populate product information when creating invoices for this commission class. You can specify more specific information per commission category and type as well as per individual commission;
+            however, if a field is unset for any of those, the information set here will be used instead. Consequently, it's required that this information be set.
+        </p>
+        @include('admin.commissions._invoice_fields', ['object' => $class, 'require' => true])
+    @endif
+
     <div class="text-right">
         {!! Form::submit($class->id ? 'Edit' : 'Create', ['class' => 'btn btn-primary']) !!}
     </div>